Fixed file paths for caffe models in compile.sh
As of version 1.4 the file paths to caffe models provided in the model zoo has changed to putting the deploy.prototxt and deploy.caffemodel into the quantized folder. This simple typo fix will help first time users to not be confused when the script is unable to find the file path.